Many functions moved into this file (from plcore.c):
	plwarn plexit plcol0 plcol1 plrgb plrgb1 plscolbg plscol0 plgcol0
	plscmap1 plscmap1f1 plscolor

Significant changes made to operation of old routines that manipulate the
color map.  plscmap1l added to set color map 1 colors using a piece-wise
linear relationship between intensity [0,1] (cmap 1 index) and position in
HLS or RGB color space.  plscmap0n added to determine the number of colors
to use in cmap 0 (must be called before plinit, or cmap 0 allocation).
This helps free up unused colors that can be used in cmap 1.

Color 0 (in cmap 0) now defaults to black.  This is done for consistency
with 16 color tek displays, where color 0 is the background.  You can
always change it back if you want the old way (in fact, you are encouraged
to not rely on the default color map now that there's a supported way of
changing it).

plabort() added to be almost identical to plwarn(), except for message
content.  Now used before a return on most error conditions.

svn path=/trunk/; revision=731