Massive reorganization and cleaning up.  Main result is that the plplot "megawidget" acts much more like a normal Tk widget.  Still not configurable; this will require extending it using itcl.  But now, creation and mapping is much more straightforward, both directly (from plserver or an extended wish) and from the plplot/tk driver.  You can do simply:
    plxframe .plw
    pack append . .plw {bottom fill expand}

and you get a plframe "megawidget", complete with plot menu (with dump,
zoom, etc) and status label.  Support widgets relevant for the plplot/tk
driver do not come up unless invoked from the tk driver (by specifying
a client to connect to).  The main drawback at this point with this method
is that direct plotting commands in Tcl must be specified using the syntax
(for the above example) ".plw.plwin cmd <command> <args>", whereas in
itcl is simply ".plw <command> <args>".

svn path=/trunk/; revision=957