This is a commit in the SHADE-EXCLUSION branch.  The modifications here allow a smooth boundary to be drawn when exclusions are specified for the plshade routines, instead of the jagged boundary we currently have in demo c16c.c.
The API of plshade has been changed, in particular the new type of the
"defined" parameter, which is now:

    PLINT (*defined) (PLFLT, PLFLT)

instead of:

    char *defined

This breaks the C++ bindings, besides causing a potential breakage of old
user code.  Furthermore, I disabled the other exclusion mechanism, based on
the values of sh_min and sh_max.  This also can have some undesirable
effects for the users.

This changes will be sitting in this branch until either Maurice or me will
have time to get back to the breakage issues.

svn path=/branches/SHADE-EXCLUSION/; revision=2735