Heavy hacking for introducing a dichotomy between static and dynamic drivers.  It all starts with a new option: --enable-dynamic-drivers
Then we introduce an assoc list of drivers that are known to be
dynamic-capable.  (This is prospective at this point, we're just
picking on null, plmeta and ps for now, since they shouldn't have
fancy linkage considerations.)

Next we introduce an assoc list of devices and the drivers that
support them.

Now hack PL_ADD_DRIVER so that when we are asked to add a driver, we
check to see if it is believed to be dynamic-capable, and if dynamic
drivers are enabled.  Also, pick out the driver needed to support the
device, and then finally add it to either the static or dynamic driver
list, as appropriate.

Finally, do the AC_SUBST thing, and add a little extra verbosity to
the config summary so we can check what's going on.

svn path=/branches/dyndrv_1/; revision=2877