I've never liked the double-duty compilation rule.  Changed this to have seperate rules for building shared and not-shared .o's.  Use GNU make pattern dependency style rules for specifying the proper behavior for compiling shared .o's, and then modified the shared lib target build rule accordingly.
This further solidifies our reliance on GNU make.

This needs to be reviewed.  The same techniques should be extended to
the other shared lib platforms, and in so doing, it may be that the
need for these implicit rules to come in through platform-specific
includable makefile components, can be expunged.  We should be able to
hoist these rules to a higher point, where they can be shared across
all platforms, adjusting only the macro-ized compilation rule
itself. Then we simply allow the dependency mechanism of make to do
its job.  State your dependencies, make will compile (only) what it
has to.

So, the current mods are a step in this direction, but more is
possible, and warranted.

svn path=/branches/dyndrv_1/; revision=2889