Beginning to rework plGetDev().  The new plGetDev() first needs to initialize the dispatch_table (handled by plInitDispatchTable()). Then it can select the device the user wants from those listed in dispatch table (handled by plSelectDev(), the old plGetDev()). Finally it will need to actually load the driver if the selected device is not already loaded (or static).  This part isn't done yet, but the first two parts now work.
This is enough to let PLplot present the full list of static and
dynamic drivers, culled from drivers/drivers.db.

Issue:  Note that we have lost the sequencing which used to be
provided by the initialization list order in plcore.h.  Probably we
should introduce a sequencing priority number to the device block, but
I'm not gonna get involved in that right now.  Right now I'm just
concentrating on getting the dynamically loadable drivers built,
loaded, and invoked.

svn path=/branches/dyndrv_1/; revision=2900