Add prototype for plLoadDriver whose responsibility it is to load a requested device driver if it is not already present.
Rework the static driver initializations so that they are skipped for
dynamic-loading-capable drivers, if dynamic loading is enabled.

Add structs PLLoadableDevice and PLLoadableDriver which provide fields
for tracking the available loadable devices and drivers through their
life cycle, as well as hooks loadable_device_list and
loadable_driver_list for storing what we learn about the available
drivers as we push our way through the device selection process.

svn path=/branches/dyndrv_1/; revision=2906