Flesh this out.  Keep records about the loadable devices as we scan the drivers description database.  Use this to index the device list into the driver list, since some drivers may provide multiple devices.
Implement plLoadDevice().  This is called after plSelectDev() (the old
plGetDev()), and checks to see if the dispatch table has valid
function pointers for the selected device.  If not, we have to load a
device.  Find the device spec, use its drvidx to look up the driver
spec.  If the driver isn't loaded, load it.  Once loaded, extract its
symbols for the requested device, initialize the dispatch table, and
send everyone home happy campers.  Dan Quayle would be proud of us.

svn path=/branches/dyndrv_1/; revision=2907