Initial commit of PLplot spec file for RedHat 6.2.  Other versions of RedHat or other rpm-based distributions will require different spec files (although not that many modifications of this spec file will be required for other distribution versions.)
This spec file has been used to build source and binary rpms successfully on
*one* RedHat 6.2 system.  It uses the built-in package relocatibility of
PLplot to safely allow non-root users to build the rpm's.  Nevertheless,
because of the scarcity of testing, it should still be consider betaware,
and you are advised to take the ordinary precautions (check the file lists
and scripts) before installing the binary rpm created by this package.

Note that like all spec files, this spec file is version specific.
It is only consistent with plplot-5.0.3 and will require some
modification for every subsequent version change of plplot that we want to
turn into an rpm.

svn path=/trunk/; revision=2948