This is a first try at installing the PLplot data files (*.fnt and *.map).  They are now properly being installed into $prefix/lib/$PACKAGE$VERSION, but I am wondering if "share" would not be more appropriate than "lib" here.  This is as simple as defining the Automake variable pkgdata_DATA in lib/Makefile.am.
However, I faced a lot of complications in doing this.  One of them
came from the fact that, traditionally, PLplot used the DATA_DIR
variable, both in source (src/plctrl.c) and in the Makefile.  I have
to redefine pkgdatadir in lib/Makefile.am in order to have this
working.  Also, in doing that, I ran into a series of headaches
involving autoheader.  I fixed everything for now by including the
required file acconfig.h in top_srcdir and by removing
include/plConfig.h from the project.  Also, autoheader (and a
seemingly mysterious "touch" command) have been included in
bootstrap.sh.  In order to have all this working, the top level
Makefile.am and configure.in had to be slightly changed.

Several issues are popping up now and will be soon discussed in
plplot-devel.

svn path=/branches/AM-LT/; revision=3011