This is a preliminary (albeit working) implementation of the selective building of drivers for inclusion in libplplot.  A convenience library (libpldrv.la) is built in the drivers directory and libplplot is linked against it.  Drivers are selected using --enable-<driver> options to configure, like in the current configuration scheme.  An AC_SUBSTituted variable called DRIVERS contains the names of the *.lo objects, on which libpldrv depends.
This implementation has still one problem, namely the fact that the
PL_DRIVERS_LIST variable in configure.in contains a list of the files
in the drivers directory, not the actual drivers, like in the current
configuration scheme.  For instance, "gd" appears in PL_DRIVERS_LIST
and PLD_gd gets defined in plDevs.h.  However, drivers/gd.c needs the
definition of PLD_jpeg and PLD_png.  I will work out a solution for
this problem as soon as possible.

For now, I can include some "easy" drivers in libplplotd, like ps and
xfig.  This is already a big step forward.

There are some miscellaneous changes going with this commit, like the
proper definition of M4_FLAGS (with -DDOUBLE, that was missing) and
the definition "CLEANFILES = sfstubs.f" in bindings/f77/Makefile.am.

svn path=/branches/AM-LT/; revision=3043