Extend character aspect ratio preservation to case where the overall aspect ratio is changed by the -a option (or equivalent call to plsdidev).  There is also some reorganization of the old code that preserved the character aspect ratio for the -freeaspect case with 90 deg rotations.
Note, the character aspect ratio preservation changes only occur in the
c_plinit routine. The idea is that when the overall aspect ratio is changed,
the xpmm and ypmm values are changed in a compensating way so that, for
example, all circular characters continue to display as circular (rather
than elliptical as in old code when the aspect ratio was changed).

svn path=/trunk/; revision=3118