Implement portrait option and change freeaspect documentation.
Joao, can your octave front end be set up to call plargs code directly (or
can plargs.c be modified so it can be called this way) so this separate list
of plargs options no longer needs to be maintained?  I only discovered by
accident (attempting to run plplot-test.sh with the -portrait option set for
the psc driver as a special test), that if I add an option to plargs.c, I
also have to add it to plsetopt.m.  Note others may not have been aware of
this requirement (which I hope disappears) for previous plargs.c changes so
that the two option lists may no longer be equivalent.

svn path=/trunk/; revision=3129