Centralization of the definition of the PLplot website ======================================================
The name of the PLplot website (that used to be www.plplot.org, but is now
plplot.sf.net) is defined in configure.in, with the variable PLPLOT_WEBSITE.
It is possible to change its value through the --with-plplot-website option
of configure.

The PLPLOT_WEBSITE variable is AC_SUBSTituted in both src/plplotdoc.xml.in
and bin/api2man.pl.in.  In the master Docbook file (src/plplotdoc.xml.in),
an entit PLplot-webiste is created, which allows constructions like '<a
href=http://&PLplot-website;">PLplot</a>' in an xml source file (see
src/intro.xml). The file bin/api2man.pl has been replaced by
bin/api2man.pl.in.  The former is automatically generated by configure.

svn path=/trunk/; revision=3153