Modified the colour routines so they now behave more like what plplot would expect them to. For example, before the driver used 256 colours whether it was asked for or not; the plplot default was for 128 colours. Now the driver will only allocate what plplot asks for, and NOT what the driver is capiable of. Things don't always look as "nice" but it is probably more "correct" to do it this way.
Adopted a slightly simplier colour allocation strategy similar to the gd
driver.

Fixed a bug (in grx20 ?) which caused two colours to be taken "out of the
available colour pool" whenever one was requested.

Modified the line drawing and filling routines, so the plplot page used by
the gnusvga driver is now "huge" (32768 pixels). This means that the bug in
the 3D code is now side-stepped, and hidden-line removal will now work
properly.

svn path=/trunk/; revision=3156