Merged dyndrv_1 branch to HEAD.  More detailed descriptions of all the mini steps were contained in the commit messages along that branch. A brief summary of the major points:
1) configure has been modified to synthesize the set of driver modules
to compile.  Each driver is compiled either as a static module, or a
dynamic module.  Separate lists hold drivers in each category.  Only
the selected driver modules are compiled.  Disabled driver modules are
not compiled.

2) Each driver may provide one or more devices.

3) The dispatch table is dynamically initialized, first with the list
of static drivers (and devices), and then with the set of dynamic
drivers.  Then this list is sorted, based on a presentation order
field in the dispatch entry data structure.

4) PLStream now has a member for the associated device.  The global
dispatch table is abolsihed, and instead the core routines vector to
the device routines by bouncing throught the dispatch table associated
with the stream.

5) Each device has a new required function which is responsible for
initializing the per-device dispatch table with the proper function
addresses.  Only this dispatch_init function is looked up using the
dynamic symbol search stuff.  This function must have a stylized name
so that it can be programmatically located, but after that, the names
of the device-specific entry points are arbitrary (provided they have
the proper signature).

svn path=/trunk/; revision=3161