First attempt to enable support for a Java binding for PLplot. Activate via --enable-java, and make sure you have JAVA_HOME set to the root of your JDK.
This much work on the configuration system is enough to get as far as
building the embrionic Java support in a Javoid (Javonian?) directory
structure.  IIRC, Java requires files to be laid out on disk in the
structure which the package import system will see.  Mapping this onto
the PLplot language porting conventions in CVS, is done, as usual,
with soft links.

This should be regarded as an initial point of discussion, as the
construction of a top-rate Java binding looks to be a substantial
undertaking.

svn path=/trunk/; revision=3165