Add some more native method wrappers.
Embellish init() so that it fetches the PLplot stream id after it
calls plinit().  This is then stored in the stream_id member field of
the PLStream Java class.  Finally, introduce a macro which is to be
expanded inside each native method wrapper, which fetches the
stream_id member from the Java class, and sets the PLplot stream id.

svn path=/trunk/; revision=3182