Implement numerous additional API bindings, especially those related to 3-d drawings for x08.java, but also several others that caught my fancy along the way.
Also left in comments for more API entry points that need to be bound
(but not all).  This is a long process, oweing to the sheer number of
functions in PLplot, and to the need to overload for float and double
in Java.

plot3d and plotsh3d are fully implemented now, including all the JNI
hocus pocus for extracting the 2-d arrays.  However, there is some
remaining issue with the "release" mechanism for the 2-d data array.
Currently we leave the arrays unreleased, which is probably a leak,
but at least it doesn't core.

svn path=/trunk/; revision=3256