For plssub calls to plP_eop() and plP_bop(). This drops page advance functionality from plssub as discussed on the list.
Added call to plP_subpInit() so that character size (which depends
on number of sub-pages) is adjusted properly.

These are tentative changes which work with the current c, tcl, and python
examples from cvs.  However, those examples only seem to work without
cross-talk between them in the tcl and python cases if the examples are
started with pladv or equivalent (see change to x02.tcl and xw02.py).  I
don't really understand this requirement. In future we may find a better way
to handle restoring the default subpages without messing with sub-page
advance.

svn path=/trunk/; revision=3313