Now the individual example script files look for the existence of global variables "device" and "file" and use them if defined. By default, "xwin" is used.
To use the Octave examples the same way as the pythomdemos or tcldemos, however,
plinit() and plend() should be conditionaly executed in each script, e.g.,

if (!exist("device"))
   plinit
end

The same for plend().

This way, the script files can be used by itself or called from a main octavedemo script file
that does the plinit() and plend().

svn path=/trunk/; revision=3334