Remove declaration of plimage, as it is now in plplot.h
Show the relationship between the xmin, max, ymin, ymax arguments of
plimage() and the similar arguments of plenv(), as well as the

Also show who to use the Dxmin, Dmax, Dymin, Dymax arguments of
plimage().

Conditionally enter plxormod() to interatively select and zoom the
image.

Kind of (!) solves the interation between pladv() and GetCursorCmd().
This is still a problem! Please try to solve it!

svn path=/trunk/; revision=3346