Olof Svensson and Alexandre Gobbo's patch to make plplot work under Windows 98, NT and 2000.  One nice thing I noticed about the patch was the plplib.mak and especially the Examples.mak files are much smaller than previously.  Here is excerpts from their description:
plplib.mak: Automatically generated.

Examples.mak: Handcrafted.

win3.cpp: Many changes...

The driver works under Windows 98, NT and 2000 for all examples except
x14c and x17c which are therefore not built by the Examples.mak file.
The driver now handles mouse events and XOR mode so example x20c works
(but is slow). What's missing in the driver is to make it a fast image
device , to implement some missing PLESC functions and to make x17c
work. I also have to make the makefile for plplot.dll.

To make the win3 device work with pyqt, we found an elegant way of
passing the windows handle without adding any new functions - we passed
it as a driver option. The win3 driver will not only work with pyqt but
with any Windows program that can provide a windows handle, for example
you could think of having a plplot window as an integral part of a
windows application.

svn path=/trunk/; revision=3381