Change limit slightly to avoid label flips at 90 and 270 degrees for slightly different floating point results.
Result was not in perfect agreement with x03c.  The "270" label moves
slightly compared to the x03c result, and this difference is consistent with
the diff result of some small systematic changes in results at just one part
of the postscript file.  I suspect this problem is the result of some
floating point logic in the plplot plptex core routine so that a small
change in the input floating point number from java makes a small but
systematic change in the position of the output 270 label.

svn path=/trunk/; revision=3384