Make ninth C example consistent with python example.  (No real change to python example except some commented out print statements.)
There are some tiny visual differences between the two results for the ninth
and tenth pages which correspond to 1 unit changes in the postscript file. I
don't really understand the origin of these differences, but I suspect we
are the victim of a nasty coincidence so that some of the pen strokes for
this example pass virtually exactly between two pixels, and the decision
about which pixel is painted rests on the last few bits of double precision
which are different in the two front ends.  I did print out a variety of
values between the two front ends.  They are consistent to 16 figures.  I
also did a lot of code inspection, and I am reasonably confident that the
two front ends are floating-point consistent to this level.  It is just for
certain examples that even this level of consistency is not good enough to
produce the same postscript plot because of the above "nasty coincidence"
argument.

svn path=/trunk/; revision=3427