Make this example more consistent with x15c.
Some rearrangement of code plus actual changes in definition of x and y
to make them consistent with x15c.

Results were identical to results from x15c, if first pattern of second page
was dropped from both.  For this pattern there were some small differences
that could be seen visually between the two results.  I have no idea what
the source of this inconsistency may be, but it seems to lie with the
x15.tcl code or the tcl implementation (perhaps the matrix implementation,
Maurice?).  When I do similar comparisons between the C and yplot results,
the two 15th examples are identical.

svn path=/trunk/; revision=3448