plcont: fix bad free memory logic
plshade: Follow all wrap fixes that were done before for plcont.  Unwrapped
plshade plots (e.g., x15.tcl) are identical, but the wrap stuff has not been
tested yet.  Plan is to test wrap logic in x16.tcl at first, then convert
that example over to much more convenient plshades API (and of course test
that wrap logic works in that case as well).

plshades: First cut defining this API with wrap logic (and a lot more)
following what was done for plshade. It compiles, but it is completely
untested!

svn path=/trunk/; revision=3664