Make this example (using nowrap approach) give consistent results with the equivalent python example, xw09.py.
Also work around long-standing clevelpos and clevelneg problem where their
size was not correct.  The workaround is to allocate the correct size in
advance (an extremely clumsy and error-prone method since the number of
positive and negative contours depends very much on the data range), but the
proper solution is to use a dynamically allocated array.  I think those are
probably called vectors in java, but I know no more than that.

Geoffrey, can you fix this properly?  Look for "Geoffrey" in comments!

This nowrap approach gives identical results to python example, but the
plan is to change to the wrap approach as soon as wrap is debugged in
javabind.c.

svn path=/trunk/; revision=3668