First commit of real code for this example.
Because plshades API is not available yet, this example is done with
pls.cont instead.  I am committing this example because it works, but
I will replace pls.cont calls shortly with pls.shade calls, and eventually
(once pls.shades is implemented in javabind.c) with pls.shades in analogy
with the rest of the 16th examples.

svn path=/trunk/; revision=3673