Take advantage of new python plpoly3 optional "counter-clockwise" flag API to produce identical results (for the first time) with other 18th examples such as x18.tcl.
Without the new counter-clockwise flag API, the old method of reversing the
arrays sort of worked, but the actual drawing of each polygon was done in a
reverse order so the details of the postscript output, for example, could
not be diffed against postscript plots produced by other front ends.
Furthermore, (and this is probably due to a plplot core bug), the hidden
line algorithm didn't work quite the same so the appearance of some of the
plots was different for the case where the line was near the edge of the
visible region.

svn path=/trunk/; revision=3678