Changed API of plpoly3 so that negative dimension n has no special significance (except you will have less than the minimum 3 points), and there is an additional argument ifcc.  If that additional argument is 1 (or greater) the points are treated as if they are layed out in counter-clockwise order.  If 0 (or less) the points are treated as if they are layed out in clockwise order.
Also changed the c++, fortran, octave, python, and tcl bindings and
the C, java, and tcl examples appropriately.  (The java binding is still
undefined for plpoly3 so no changes there, and the python example already
used this form of argument list so no changes there.  The octave example
has a different way of doing things (last argument of +1 or -1 corresponds
respectively to ifcc of 0 and 1) which I preserved, but Joao may want to
change this to use the native plpoly3 rather than his special version.)

The C, tcl, python, and octave 18th examples (which all call plpoly3
directly or indirectly) give identical results which is a good test that
the API changes have been done in a consistent manner.

svn path=/trunk/; revision=3703