Configuration changes to support new names for plplot python modules, and also a swig-based approach for generating the plplotcmodule.c wrapper file.
plplot (plplot.py) is the user-friendly module that python users will see.
This is the equivalent of the old pl module.  plplot wraps the raw python
extension module plplotc, which implements in python the exact equivalent of
the common C API for the plplot library except that the redundant array
dimensions are dropped.  The pyqt_pl module (and its equivalent wrapper C
code, pyqt_plmodule.c) has been renamed plplot_widget since the functions
contained in this extension module can be used by pyqt as well as other
python widgets such as Tkinter.

svn path=/trunk/; revision=3714