AWI for Gary Bishop.
Now make calls from shade_triangle (called by new plotsh3d) to plP_fill
through the clipping interface. (Gives much more meaningful result if the
plot is poorly scaled and thus there is a better chance that the user will
recognize the scaling problem and solve it.)

Tweak plotsh3d documentation.

Refine the #if WANT_MISSING_TRIANGLES (heh) logic that define the old code
that is going to be removed after a while.

svn path=/trunk/; revision=3723