plcont: put in check for correct number of arguments
plshades: interface is completely redone to do fairly rigorous
self-consistency checks on the argument list as well as to provide many more
variations of arguments than before.  Also, I did fairly extensive checks of
the new interface using example 16 (which exercises plshades).  Not only the
old version of xw16.py, but also the new version (which takes advantage of
the defaults allowed here) produced identical results to the tcl results for
example 16.

svn path=/trunk/; revision=3743