Tighten up checks on yg and number of arguments for plcont and plshades.
New user-friendly interface for plshade (mostly following that of plshades).
This is only lightly tested because the 15th example is so limited in its
use of plshade, but that result was good (result the same as 15th example
from tcl), and this is probably adequate testing because of the
comprehensive plshades argument logic done before.

svn path=/trunk/; revision=3745