Revert pleop "fix" at end which was meant to stop cross-talk between restoration of default cmap1 work at end of example 8 and the colours of the last page of that example (which should remain gray scale rather than default).  The analogous fix for the tcl situation (which I did not commit) resulted in a segfault so that's why I am reverting this.
After this reversion the last page of the 8th python and tcl examples
will have default colours (in error) rather than the specified grey scale
for the png driver.  I have to think a bit more deeply whether this is a bug
or whether it is a "feature" that we have to work around somehow.

svn path=/trunk/; revision=3788