Configuration support for linking libplplot against the tk, xwin, tkwin dynamic drivers.  Note this scheme should be easily modified to the case when we split off the tcl stuff from libplplot, link libplplot against libtclplplot, and link libtclplplot against the tk, xwin, and tkwin drivers.
The pkg*.in changes and addition conditionally define the appropriate
shared object (dll) and link symbols for the Makefile.

The initlib.in changes define the Makefile symbols PLLIB_LIBRARIES and
PLLIB_LINKS in terms of the above defined symbols and also the corresponding
symbols for libtclmatrix.

The lib_sh_linux.in changes replace the libtclmatrix symbols by
the corresponding PLLIB_LIBRARIES and PLLIB_LINKS symbols.

Eventually this change should be propagated to the other lib_sh*.in files by
those who know the linking mysteries on those various platforms.  For now
those other platforms should work as is if you don't specify dynamic
drivers.

svn path=/trunk/; revision=3817