New name for dynamic drivers:  drivers/*.drv ==> drivers/*_drv.so. (For one key test case valgrind was getting confused by the lack of ".so" on the ends of the driver dll's so I decided to put such a name-ending in there in case there was a similar dynamic loader assumption that was giving us the pecular results discussed below.)
Configuration support for linking libplplot against the tk, xwin, and tkwin
dynamic drivers.

In dyndrv.in commented out the link of all drivers against libplplot and
also got rid of the corresponding Makefile dependencies.  This change was
necessary to remove a circular Makefile dependency (libplplot depends on tk,
xwin, and tkwin dynamic drivers while those drivers depend in turn on
libplplot.)

This change gives no linker errors and fine results for directly loaded
examples such as x01c, but dynamically loaded examples from python (and
apparently from java also according to the warnings in the comments about
this) no longer work.  Also after this change valgrind started complaining
about the name of the dynamic driver (so I changed it as above).  None of
the results of this change make sense to me so more discussion will be
required on list to get this straightened out.

svn path=/trunk/; revision=3818