OOPS.  Forgot to modify this file so that tk, xwin, and tkwin are treated like other dynamic drivers. The net effect of this bug was
(a) double entries in menus (presumably one static and one dynamic for each
of the 3 drivers and now gone after this bug fix)

(b) tk looked like a special dynamic driver that libplplot had to link to.
(My hand linking tests indicate this is no longer necessary after this bug
fix [although libplot still has to link to xwin and tkwin as before]). BTW,
I believe Joao also said on list some time ago that tk could be dynamic with
no problem (except for the xwin linking troubles), and this is consistent
with my present results.

Anyhow, I will do full test of removing tk from the special list, and if
that test succeeds do a commit of all the associated file changes.

svn path=/trunk/; revision=3819