Move tk from the "special" dynamic driver list (which now just has xwin and tkwin) to the ordinary dynamic driver list.  The ordinary drivers link to libplplot and vice versa for the special drivers.
This completes the dynamic driver work for the plplot/tmp location, and the
results have been tested with my standard non-interactive and interactive
tests for that location.  I have also done new interactive tests
with plserver and wish both sourcing runAllDemos.tcl in that location as
well as spot-checking some java examples.  All was well.

Next on the agenda (tomorrow) is to complete the dynamic driver work for
the installed version.

svn path=/trunk/; revision=3821