Split xwin.c into two parts to sort out cross-linking problems.
xwin_common.c contains functions which are required both by xwin.c and
routines in libpltcl.

xwin.c no longer contains functions or symbols referred to by any library.

The minor plframe.c and plxwd.h change is to handle plplot_ccmap by defining
it in plxwd.h as a static variable.  (Before this was done with an extern
in plframe which then made libpltcl depend on xwin.c.)

The configuration changes are to drop the special status of the xwin driver
and to build xwin_common.c and make the resulting xwin_common.o file part of
libpltcl. With this change there are no more special xwin symlinks, libpltcl
is no longer linked to the xwin driver, etc.  Instead, the xwin driver is
linked to plplot libraries that are needed to resolve its symbols similarly
to other drivers.

svn path=/trunk/; revision=3835