Support routines for drivers interested in using freetype# rendered fonts instead of plplot plotter fonts.
Any bitmap driver should be able to use any of these freetype fonts from
plplot if these routines are properly initialised.

Freetype support is not intended to be a "feature" of the common API, but
is  implemented as a driver-specific optional extra invoked via the -drvopt
command line toggle. It is intended to be used in the context of
"PLESC_HAS_TEXT" for any bitmap drivers without native font support.
Implementing freetype in this manner minimise changes to the overall API.
Because of this approach, there is not a "wealth" of font options available
to the programmer. You can not do anything you can't do for a normal
freetype plotter font like boldface. You can do most of the things that you
can do with a plotter font however, like greek characters superscripting,
and selecting one of the four "pre-defined" plplot font types.

Freetype support is turned on/off at compile time by defining
"HAVE_FREETYPE".

[# Freetype is a well known opensource library that supports a gerth of
font formats including TrueType, OpenType, Adobe Type1, Type42 etc... the
list seems almost endless.]

svn path=/trunk/; revision=3889