Reorder plD_tidy_png so that it will compile with gcc.
move call to init_freetype_lv2 from init part of code to bop part of code
(just after GD image is created).  Otherwise, in -drvopt text,smooth mode
you will get a call to plscmap0 and therefore a call to use the GD colours
when no GD image is defined ==> segfault.

Note, this fix is not quite right yet.  x07c now goes through without a
segfault in familied -drvopt text,smooth mode, but the antialiasing looks
bad after the first page.  So it is simply a basis for discussion with Andrew,
and by no means the final fix.

svn path=/trunk/; revision=3901