Restore "minor" changes in title, title colour, and order of plplot commands to previous values.
I am reverting this part of the previous change because I don't think it
made any obvious improvement (in fact I preferred the previous versions in
any case), and also it is extremely important from the testing perspective
to maintain consistency as much as possible with example 8 for other
interfaces (once those other example 8's are upgraded to also use plotfc3d).

That said, I am willing to change all other python, tcl, java, and yorick
examples to be consistent with any *important* improvement to C examples
(such as plotfc3d).  But it is a fair amount of work to keep this
equivalence of the examples maintained so please be careful of what you do
change in the C examples.

svn path=/trunk/; revision=3925