Make most dynamic drivers work under AT.
For example, plmeta now works, and plrender also works fine for the plmeta
file that was created.

There are still some problem dynamic drivers.  I still need to add the
correct library lines to build gd_drv.so so that the png and jpeg devices
work.  plserver works fine, and so does loading tk_drv.so and running
examples under wish. But -dev tk (which uses the same shared object) does
not work for some reason which I am still tracking down.

svn path=/branches/AT/; revision=4092